Refractor Async<T> into a new file
This commit moves the implementation for Async<T> into a new file called
io.rs. This is done, as adding wasm32-unknown-unknown does not support async
I/O, so it's easiest to move it all to another file to cut if off.

As this commit will make up the majority of the lines of the overall PR, I've
elected to make it separate for ease of review.
